聊天功能实现

+++ chatgpt-client/src/main/java/com/szmengran/chatgpt/dto/Usage.java
@@ -12,17 +12,17 @@ public class Usage {
      * The number of prompt tokens used.
      */
     @JsonProperty("prompt_tokens")
-    long promptTokens;
+    int promptTokens;
 
     /**
      * The number of completion tokens used.
      */
     @JsonProperty("completion_tokens")
-    long completionTokens;
+    int completionTokens;
 
     /**
      * The number of total tokens used
      */
     @JsonProperty("total_tokens")
-    long totalTokens;
+    int totalTokens;
 }
